[PORT FROM MAIN]FRI State Machine update problem fix
BZ: 15851

When configuring polling loop, NFC JNI layer will try to update the FRI
state machine. If a state machine is in a "transition" state when trying
to be updated, the error will be generated which provokes abort and
restart of NFC service. That is why a proper error return code
(NFCSTATUS_BUSY) has to be provided to upper layers from libnfc. On the
other side, the NFC JNI will handle this kind of error with its retry
mechanism. If after 3 retries it does not succeed to configure polling
loop, then it shall call the abort.

Change-Id: I430dd70457cfd44cf4ac6a0dd4c90a6fa9db61e3
Orig-Change-Id: I1ceecfdc161c9a47338c16c585517df86e72771b
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/26744
Reviewed-by: Champciaux, NicolasX <nicolasx.champciaux@intel.com>
Tested-by: Champciaux, NicolasX <nicolasx.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Ken Wahid <kenx.wahid@intel.com>
Reviewed-on: http://android.intel.com:8080/30834
Reviewed-on: http://android.intel.com:8080/58944
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/76106
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>