[PORT FROM MAIN][NFC]FW download status fix
BZ: 21973

In case we have FW download failure, the wrong status (OK) is provided
to Java layer. This causes a problem, when trying to enable NFC
afterwards or it can cause problems in operational NFC mode. To avoid
this, the correct status shall be provided in order to trigger the retry
machanism of FW download.

Category: Aosp improvement
Domain: CWS.NFC-HCI
Origin: internal
Upstream-Candidate: yes

Change-Id: I5e6e5d6546aa8c3ed69dfbb7285be6ebef78d427
Orig-Change-Id: I5fc9189f93ddf3e1e9ece16e3457f3fe1b7d29f4
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/75801
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Reviewed-by: Wahid, KenX <kenx.wahid@intel.com>
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-on: http://android.intel.com:8080/122762
Reviewed-by: Deprez, Olivier <olivier.deprez@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>