[PORT FROM MAIN][NFC]Proper termination of Ndef and Snep server threads when disabling NFC
BZ: 12338 25360 90711 95738

Ndef and Snep server threads are left hanging every time we do disable
of NFC. After enabling of NFC again, new threads are created, but the
previous ones also remain. This leads to thread leakage if doing NFC
enable/disable plenty of times. That's why the new server socket flags
are introduced, to keep tracking when the NDEF or SNEP server socket is
closing. When this is the case, a proper cleanup is done and the threads
are terminated properly.

Category: Aosp improvement
Domain: CWS.NFC-HCI
Origin: internal
Upstream-Candidate: yes

Change-Id: I197541c44e8f87cdfd7c93807e9fafcf9526116f
Orig-Change-Id: Ic4a3de9abebbad7458a92c5de95b64a1c1cd6d1f
Orig-Change-Id: I330b1c2f45a58c669e86dce03fbad349f716c8f6
Orig-Change-Id: I781eadb9697a1cb12ce9ca5d85d8cfcda564aeed
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Signed-off-by: Ken Wahid <kenx.wahid@intel.com>
Reviewed-on: http://android.intel.com:8080/75800
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Reviewed-by: Wahid, KenX <kenx.wahid@intel.com>
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-on: http://android.intel.com:8080/122660
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>