[PORT FROM MAIN]Link status callback conditioned with link status changed
BZ: 36010

The execution body of link status callback should be conditioned by link
status changed from activated to deactivated state and inverse. This
will avoid crash if the same callback is called twice in a row without
setting a context data in between with phLibNfc_Llcp_CheckLlcp function.
When callback is called twice in a row, the first call will free at the
end the memory allocated for context data, where a second callback will
use the same memory area. That will crash nfc process. The link status
callback is called second time to notify the upper layers that there was
a forced disconnection from remote NFC device while the transaction was
ongoing.

Change-Id: I3fe84152584a8f51a93fd41320a7d91b9fe5710b
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/49028
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/58957
Reviewed-by: Wahid, KenX <kenx.wahid@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/76199
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>