[PORT FROM MAIN] Fix UICC <> CLF session setup
BZ: 46879

DO NOT UPSTREAM

In some cases creation of the session between the CLF and the UICC
can be interrupted because NFC controller boots with the SWP link activated
but later on NFC stack de-activates SWP if card emulation through the UICC
was not selected.

To ensure the session is setup correctly we add 1 sec. delay time
before we de-activate the SWP link.

Change-Id: I7e9317814a07c42c24cdedf2590014420273ed17
Orig-Change-Id: Ic5a9e48b6ae9a1f3e640548b7c4366f095753e80
Signed-off-by: Sebastian Niciarz <sebastian.niciarz@intel.com>
Reviewed-on: http://android.intel.com:8080/63870
Reviewed-by: Rebraca, DejanX <dejanx.rebraca@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Sebastian Niciarz <sebastian.niciarz@intel.com>
Reviewed-on: http://android.intel.com:8080/79495
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>