Modifications needed to use NfcService with both NCI and HCI stack
BZ: 90332

DO NOT UPSTREAM

For a successful compilation we need to extend the DeviceHost I/F
to support both NCI and modified by Intel HCI stack (NXP SE patch).

We also added the implementation of the NFC modem manager client
in the NCI stack.

To be squashed with SE patch from NXP.

Change-Id: Idc4d5dcd1179f768a64cc4728296927a113c25e4
Signed-off-by: Pierre Appietto <pierrex.appietto@intel.com>
Signed-off-by: Sebastian Niciarz <sebastian.niciarz@intel.com>
Reviewed-on: http://android.intel.com:8080/96796
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Rebraca, DejanX <dejanx.rebraca@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>