Fix ANR during Factory Reset
BZ: 90044

MASTER_CLEAR_NOTIFICATION intent is received in the context of the main
thread. On receive of this intent, the Nfc service will try to execute
EE wipe, which means to temporarily enable NFC if NFC was disabled for
EE wipe execution. During that time, the main thread is blocked, waiting
for the end of these actions. On the other side, modem manager callbacks
registration is also done in the context of the main thread. Since the
main thread is blocked, this registration will not be performed and NFC
cannot be enabled. That's will cause an ANR in MASTER_CLEAR_NOTIFICATION
broadcast.
Here, the fix is to remove BroadcastReceiver for
MASTER_CLEAR_NOTIFICATION from to the context of the main thread to
dedicated thread, which will allowed modem manager callbacks to be
registered properly.

Change-Id: I944859ef93face52d3439d6745cef264116da7e9
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Reviewed-on: http://android.intel.com:8080/94602
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>