[PORT FROM MAIN][BT][NFC] Fix BT pairing on tablets wo/ HSP support
BZ: 146901

Current implementation of NFC-BT handover does not handle platforms which
would not support both HSP and A2DP.
In such scenario, the NFC app is blocked in its internal state machine
as it cannot retrieve the BT Headset proxy.

This patch starts a 1s timer upon connection to Headset or A2DP proxies.
Upon timer timeout, it is assumed that any profile not available is
actually not supported and the NFC-BT handover is completed in this
configuration.

Change-Id: I8caf23613e0e07fcab08c5a0e5665afe04b7cd45
Orig-Change-Id: I39dc51a3b59dc85209a4dafb209556ac6253fa27
Category: device enablement
Domain: CWS.BT-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Patrick Combes <patrick.combes@intel.com>