[PORT FROM MAIN JB-MR2][NFC] Terminate Ndef/Snep threads on NFC disable
BZ: 12338 25360 90711 95738

Ndef and Snep server threads are left hanging every time we do disable
of NFC. After enabling of NFC again, new threads are created, but the
previous ones also remain. This leads to thread leakage if doing NFC
enable/disable plenty of times. That's why the new server socket flags
are introduced, to keep tracking when the NDEF or SNEP server socket is
closing. When this is the case, a proper cleanup is done and the threads
are terminated properly.

Change-Id: I941127c0dbc379be762ba14c25ad43d1ac8de5bc
Orig-Change-Id: I197541c44e8f87cdfd7c93807e9fafcf9526116f
Category: device enablement
Domain: CWS.NFC-PN544
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Signed-off-by: Ken Wahid <kenx.wahid@intel.com>
Signed-off-by: Olivier Deprez <olivier.deprez@intel.com>
Reviewed-on: http://android.intel.com:8080/75800
Reviewed-on: http://android.intel.com:8080/122660
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Reviewed-by: Wahid, KenX <kenx.wahid@intel.com>
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Tested-by: cactus <cactus@intel.com>