[PORT FROM MAIN JB-MR2][NFC] FW download status fix
BZ: 21973

In case we have FW download failure, the wrong status (OK) is provided
to Java layer. This causes a problem, when trying to enable NFC
afterwards or it can cause problems in operational NFC mode. To avoid
this, the correct status shall be provided in order to trigger the retry
machanism of FW download.

Change-Id: Ib49dfeb8a6a253a6f4c02387d9bb06d41f150038
Orig-Change-Id: I5e6e5d6546aa8c3ed69dfbb7285be6ebef78d427
Category: device enablement
Domain: CWS.NFC-PN544
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Signed-off-by: Dejan REBRACA <dejanx.rebraca@intel.com>
Signed-off-by: Olivier Deprez <olivier.deprez@intel.com>
Reviewed-on: http://android.intel.com:8080/75801
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Reviewed-by: Wahid, KenX <kenx.wahid@intel.com>
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>