Allow the user to enable NFC in APM
BZ: 156193 114043

This patch is fixing several NFC Issues we had in APM:
The CSM is seeing the modem as "Not Available" while
APM is on, which was causing a reset of the NFC service.

The SIM detection machine state had to be optimized for
APM, in order to avoid useless reset of the stack.

Change-Id: I1292616029c3cf08309f4b2d4e4c775d80b07552
Category: device enablement
Domain: CWS.NFC-PN547
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Pierre Appietto <pierrex.appietto@intel.com>