CE: Make HCE able to handle long APDU (>250 bytes)
BZ: 163352

Due to a bug in the type error  in the prototype of
RoutingManager::handle() function Android cannot handles
APDUs which have size above 255 bytes.
In order to be able to handle longer APDUs, the size has  to be
stored within a UINT16.

Change-Id: I9c940e425218bed7a4a2f7f0173d84588555d5de
Category: device enablement
Domain: CWS.NFC-CE
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>